Extract arrow codec
