+++ pkg/loki/modules.go
@@ -1558,14 +1558,6 @@ func (t *Loki) initV2QueryEngineWorker() (services.Service, error) {
 		StreamFilterer: t.Cfg.QueryEngine.Executor.StreamFilterer,
 	}
 
-	if cache.IsCacheConfigured(t.Cfg.QueryEngine.TasksResultCache.CacheConfig) {
-		tasksCache, err := cache.New(t.Cfg.QueryEngine.TasksResultCache.CacheConfig, prometheus.DefaultRegisterer, logger, stats.ResultCache, constants.Loki)
-		if err != nil {
-			return nil, fmt.Errorf("creating task results cache: %w", err)
-		}
-		workerParams.Executor.TaskCache = tasksCache
-	}
-
 	worker, err := engine_v2.NewWorker(workerParams)
 	if err != nil {
 		return nil, err
